feat: add --run to llama stack build
--run runs the stack that was just build using the same arguments during the build process (image-name, type, etc)

This simplifies the workflow a lot and makes the UX better for most local users trying to get started rather than having to match the flags of the two commands (build and then run)

Also, moved `ImageType` to distribution.utils since there were circular import errors with its old location
